Creando la funcion guardarPersona para insertar en la base de datos

+++ app/src/main/java/com/example/registrodepersonas/models/Personal.kt
@@ -6,9 +6,9 @@ import androidx.room.PrimaryKey
 @Entity
 data class Personal (
     @PrimaryKey(autoGenerate = true)
-    val idPersona: Int,
+    val idPersona: Long,
     val nombre:String,
     val email:String,
-    val idOcupacion: Int,
+    val idOcupacion: Long,
     val salario:String
         )
